+++ src/ui/prompts.ts
@@ -167,7 +167,6 @@ export async function promptCsvOptionsIfNeeded(
 
 	if (looksLikeHeader) {
 		return await handleCsvHeaderBasedSelection(headerCells);
-	} else {
-		return await handleCsvIndexBasedSelection(headerCells);
 	}
+	return await handleCsvIndexBasedSelection(headerCells);
 }